requirements.txt
requests==2.31.0
pandas>=1.3.0
matplotlib==3.8.0
plotly==5.18.0
dash==2.14.0
beautifulsoup4==4.12.2
python-dotenv>=0.19.0
dune-client>=1.3.0
